# docs/swagger.yaml
basePath: /api/
definitions:
...
  db.ListTutoresByMateriaRow:
    properties:
      activo:
        type: boolean
      asignacionID:
        type: integer
      fechaAsignacion:
        $ref: '#/definitions/pgtype.Date'
      materiaID:
        type: integer
      tutorApellido:
        type: string
      tutorID:
        type: integer
      tutorNombre:
        type: string
    type: object
...
  db.Materia:
    properties:
      codigo:
        type: string
      creditos:
        type: integer
      descripcion:
        $ref: '#/definitions/pgtype.Text'
      facultad:
        type: string
      materiaID:
        type: integer
      nombre:
        type: string
    type: object
...
  db.TutorMateria:
    properties:
      activo:
        type: boolean
      asignacionID:
        type: integer
      fechaAsignacion:
        $ref: '#/definitions/pgtype.Date'
      materiaID:
        type: integer
      tutorID:
        type: integer
    type: object
...
  handler.CreateMateriaRequest:
    properties:
      codigo:
        example: MATH101
        type: string
      creditos:
        example: 4
        type: integer
      descripcion:
        example: Introducción al cálculo diferencial e integral
        type: string
      facultad:
        example: Ingeniería
        type: string
      nombre:
        example: Cálculo I
        type: string
    type: object
  handler.CreateMateriaResponse:
    properties:
      materia_id:
        type: integer
    type: object
...
  handler.CreateTutorMateriaRequest:
    properties:
      activo:
        example: true
        type: boolean
      fecha_asignacion:
        example: "2024-12-15"
        type: string
      materia_id:
        example: 1
        type: integer
      tutor_id:
        example: 1
        type: integer
    type: object
  handler.CreateTutorMateriaResponse:
    properties:
      asignacion_id:
        type: integer
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: Descriptive error message
        type: string
    type: object
...
  handler.UpdateMateriaRequest:
    properties:
      codigo:
        example: MATH101
        type: string
      creditos:
        example: 4
        type: integer
      descripcion:
        example: Introducción al cálculo diferencial e integral
        type: string
      facultad:
        example: Ingeniería
        type: string
      nombre:
        example: Cálculo I
        type: string
    type: object
...
  handler.UpdateTutorMateriaRequest:
    properties:
      activo:
        example: false
        type: boolean
    type: object
...
  pgtype.Date:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
...
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
...
host: matwa.tail013c29.ts.net
info:
  contact: {}
  description: This is the API for my project.
  termsOfService: http://swagger.io/terms/
  title: Proyecto de Ingenieria de  Datos API
  version: "1.0"
paths:
...
  /v1/materias:
    get:
      description: Retrieves materias filtered by faculty.
      parameters:
      - description: Faculty name
        in: query
        name: facultad
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved materias
          schema:
            items:
              $ref: '#/definitions/db.Materia'
            type: array
        "500":
          description: Failed to retrieve materias
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List Materias by Faculty
      tags:
      - Materias
    post:
      consumes:
      - application/json
      description: Creates a new materia (subject).
      parameters:
      - description: Materia Data
        in: body
        name: materia
        required: true
        schema:
          $ref: '#/definitions/handler.CreateMateriaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created materia
          schema:
            $ref: '#/definitions/handler.CreateMateriaResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to create materia
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create Materia
      tags:
      - Materias
  /v1/materias/{id}:
    delete:
      description: Deletes a materia by its ID.
      parameters:
      - description: Materia ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Successfully deleted materia
        "400":
          description: Invalid materia ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to delete materia
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete Materia
      tags:
      - Materias
    get:
      description: Retrieves a specific materia by its ID.
      parameters:
      - description: Materia ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved materia
          schema:
            $ref: '#/definitions/db.Materia'
        "400":
          description: Invalid materia ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Materia not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to retrieve materia
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get Materia by ID
      tags:
      - Materias
    put:
      consumes:
      - application/json
      description: Updates an existing materia.
      parameters:
      - description: Materia ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Materia Data
        in: body
        name: materia
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateMateriaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated materia
          schema:
            $ref: '#/definitions/db.Materia'
        "400":
          description: Invalid request body or materia ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Materia not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to update materia
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update Materia
      tags:
      - Materias
...
  /v1/tutor-materias:
    get:
      description: Retrieves tutores assigned to a specific materia.
      parameters:
      - description: Materia ID
        in: query
        name: materia_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved tutores
          schema:
            items:
              $ref: '#/definitions/db.ListTutoresByMateriaRow'
            type: array
        "400":
          description: Invalid materia ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to retrieve tutores
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List Tutores by Materia
      tags:
      - TutorMaterias
    post:
      consumes:
      - application/json
      description: Creates a new tutor-materia assignment.
      parameters:
      - description: TutorMateria Assignment Data
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/handler.CreateTutorMateriaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created assignment
          schema:
            $ref: '#/definitions/handler.CreateTutorMateriaResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to create assignment
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create TutorMateria Assignment
      tags:
      - TutorMaterias
  /v1/tutor-materias/{id}:
    delete:
      description: Deletes a tutor-materia assignment by its ID.
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Successfully deleted assignment
        "400":
          description: Invalid assignment ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to delete assignment
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete TutorMateria Assignment
      tags:
      - TutorMaterias
    get:
      description: Retrieves a specific tutor-materia assignment by its ID.
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved assignment
          schema:
            $ref: '#/definitions/db.TutorMateria'
        "400":
          description: Invalid assignment ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Assignment not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to retrieve assignment
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get TutorMateria by ID
      tags:
      - TutorMaterias
    put:
      consumes:
      - application/json
      description: Updates an existing tutor-materia assignment.
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Assignment Data
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateTutorMateriaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated assignment
          schema:
            $ref: '#/definitions/db.TutorMateria'
        "400":
          description: Invalid request body or assignment ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Assignment not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to update assignment
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update TutorMateria Assignment
      tags:
      - TutorMaterias
...
schemes:
- https
swagger: "2.0"